+++ CHANGELOG.md
@@ -3,5 +3,6 @@
 ### Added
 
 ### Changed
+- Structured code / improved logging for better diff'ing
 
 ### Fixed